+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Service order possibility (@mkasztelnik)
 - Configure `database_cleaner` for rspec (@mkasztelnik)
 - Service terms of use with markdown renderer (@mkasztelnik)
+- Send email to order owner when order is created or updated (@mkasztelnik)
 
 ### Changed